+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Select menu overflow
-- `Tooltip` position calculation. Use container instead of window to calculate collisions.
+- `Tooltip` position calculation. Use container and window to calculate collisions.
 
 ## [9.124.0] - 2020-07-02